requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
attrs==21.4.0
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
iniconfig==1.1.1
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
PyPDF2==1.26.0
pytest==6.2.5
pytest-django==4.5.2
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
webencodings==0.5.1
xhtml2pdf==0.2.5
